% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_output.R
\name{chk_output}
\alias{chk_output}
\title{Check if out is all of a class}
\usage{
chk_output(arg_list, l, type)
}
\arguments{
\item{arg_list}{Master argument list}

\item{l}{List to check}

\item{type}{Class to check}
}
\value{
Logical true if all objects are class type, false if any are not or are missing
}
\description{
Check if out is all of a class
}
