% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_output_future.R
\name{chk_output_future}
\alias{chk_output_future}
\title{Check if future list is all of a class}
\usage{
chk_output_future(arg_list, l, type)
}
\arguments{
\item{arg_list}{Master argument list}

\item{l}{List to check}

\item{type}{Class to check for}
}
\value{
Logical, true if all objects are class type, false if any are not or are missing
}
\description{
Check if future list is all of a class
}
