% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_future_list.R
\name{create_future_list}
\alias{create_future_list}
\title{Create a future list object}
\usage{
create_future_list(arg_list)
}
\arguments{
\item{arg_list}{Master arg list}
}
\value{
A nested list with time-intervals and scenarios
}
\description{
Create a future list object
}
