% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_dir.R
\name{create_model_dir}
\alias{create_model_dir}
\title{Create a directory for modelling}
\usage{
create_model_dir(path)
}
\arguments{
\item{path}{The path where the directory should be created}
}
\value{
Nothing, side effect is a directory is create with arg_list.yaml
and pipeline.R, and a directory for adding bioclimatic rasters.
}
\description{
Create a directory for modelling
}
