% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spp_occurence_sf_list.R
\name{create_spp_occurence_sf_list}
\alias{create_spp_occurence_sf_list}
\title{Creat species occurence sf::sfc list}
\usage{
create_spp_occurence_sf_list(spp_list, bb, creds,
  path = "species-occurence-tables", start_year, end_year,
  min_ocurences = 50, keep_top_n = F, n_to_keep = 5)
}
\arguments{
\item{spp_list}{List of scientific species or genus names}

\item{bb}{sf::bbox object like, used to crop extent of occurences}

\item{creds}{GBIF user credentials}

\item{start_year}{Start year for occurence query}

\item{end_year}{End year for occurence query}
}
\value{
A named list of sf::sf data.frame(s) per species name.
}
\description{
Creat species occurence sf::sfc list
}
