% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spp_scenario.R
\name{plot_spp_scenario}
\alias{plot_spp_scenario}
\title{Plot historical, future and the difference in occurence probability}
\usage{
plot_spp_scenario(spp_out, species, stat = "EMca", year = 2090,
  scenario = "rcp85", crs = "mercator", iso3 = NULL)
}
\arguments{
\item{spp_out}{A sf::sfc object with spp occurence probability}

\item{species}{Species to plot}

\item{stat}{The statistic to plot}

\item{year}{The future time-interval to plot}

\item{scenario}{The future scenario to plot}

\item{crs}{Choose a map projection}

\item{iso3}{Country iso alpha 3 code}
}
\value{
Nothing, side effect is a plot
}
\description{
Plot historical, future and the difference in occurence probability
}
