% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_species_names.R
\name{select_species_names}
\alias{select_species_names}
\title{Select species names by country iso3 code}
\usage{
select_species_names(path, iso3, field = "canonicalName")
}
\arguments{
\item{path}{Path to csv with columns `iso3` and the defined field}

\item{iso3}{ISO alpha 3 code for country}

\item{field}{The name of the field in the csv with the name}
}
\value{
Vector string of names
}
\description{
Select species names by country iso3 code
}
\examples{
arg_list = yaml::read_yaml("arg_list.yaml")
path = do.call(file.path, as.list(arg_list$search_spp_list_path))
iso3 = 'TZA'
spp_list <- select_species_names(path, iso3)
}
